std::vector<uint16_t> getIndices() -> std::vector<uint32_t> getIndices()

+++ render/core/draw_data.cpp
@@ -22,8 +22,8 @@ std::vector<Vertex> getVertices() {
   return vertices;
 }
 
-std::vector<uint16_t> getIndices() {
-  const std::vector<uint16_t> indices = {
+std::vector<uint32_t> getIndices() {
+  const std::vector<uint32_t> indices = {
 
 
       4, 5, 7, 5, 6, 7,
